requirements.txt
matplotlib
pandas<2.0.0
pypistats
python-dotenv
requests
seaborn
